Time: 15 ms (10.71%) | Memory: 24.9 MB (5.10%) - LeetSync

+++ 136-single-number/single-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int singleNumber(vector<int>& nums) {
         int ans;
-        unordered_map<int,int>hash;
+        unordered_map<long long,int>hash;
         for(int i = 0;i<nums.size();i++){
             hash[nums[i]]++;
         }
